+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-### 2.2.0 (Next)
+### 2.2.0
 - Rename former `sort_list` to `sort_list_string`.
 - Update `repeat` function to also accept string input.
 - Add `combinations`, `count`, and `sort_list_number` functions.